fixed confirming twice while editing page title
